+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [ENHANCEMENT] Distributor: Add count, spans, and buckets validations for native histogram. #7072
 * [BUGFIX] Compactor: Avoid race condition which allow a grouper to not compact all partitions. #7082
 * [BUGFIX] Fix bug where validating metric names uses the wrong validation logic. #7086
+* [BUGFIX] Ring: Change DynamoDB KV to retry indefinetly for WatchKey. #7088
 
 ## 1.20.0 in progress